# nodejs-server-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a the API Specification for the call center developed by TXM."
  version: "0.0.1"
  title: "API Aplication Sip Caster Performance"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "marellano@txmglobal.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:3030"
basePath: "/callCenter/api/v0"
...
schemes:
- "http"
paths:
...
  /AverageCallsTime/getCallsbyAverage:
    post:
      tags:
      - "AverageCallsTime"
      summary: "Average Time Answered Calls"
      description: "This service returns average time answered calls and answered\
        \ calls average waiting time"
      operationId: "getCallsbyAverage"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "Service"
        in: "formData"
        description: "Name of the service"
        required: true
        type: "string"
      - name: "Day"
        in: "formData"
        description: "Format YYYY-MM-DD"
        required: true
        type: "string"
        format: "string"
      - name: "Country"
        in: "formData"
        description: "Example: Brazil"
        required: false
        type: "string"
      responses:
        200:
          description: "Successful operation!"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/AverageCallsTime"
        400:
          description: "Error...!"
      security:
      - CallCenterCalls_auth:
        - "read:AverageCallsTimeInformation"
      x-swagger-router-controller: "AverageCallsTime"
securityDefinitions:
  CallCenterCalls_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      read:AbandonedCallInformation: "Abandoned calls information"
      read:AnsweredCallsInformation: "Answered calls information"
      read:CallsByDurationInformation: "Information of calls by duration"
      read:AverageCallsTimeInformation: "Information of calls by average"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  AverageCallsTime:
    type: "object"
    properties:
      Service:
        type: "string"
      Dat:
        type: "string"
      Country:
        type: "string"
        example: "null"
      AverageTalkTimeDuration:
        type: "integer"
      AverageWaitingTimeOfCallsAnswered:
        type: "integer"
    example:
      Service: "ACARI"
      Day: "2018-02-10"
      Country: "Jamaica"
      AverageTalkTimeDuration: 100
      AverageWaitingTimeOfCallsAnswered: 200
    xml:
      name: "Calls by duration"
...